Use generated CHARACTER_INFO

+++ Packages/RagnarokNetwork/Sources/RagnarokNetwork/Models/AccountInfo.swift
@@ -13,13 +13,13 @@ public struct AccountInfo: Sendable {
     public private(set) var accountID: UInt32
     public private(set) var loginID1: UInt32
     public private(set) var loginID2: UInt32
-    public private(set) var sex: UInt8
+    public private(set) var sex: Int
 
     init(packet: PACKET_AC_ACCEPT_LOGIN) {
         self.accountID = packet.AID
         self.loginID1 = packet.login_id1
         self.loginID2 = packet.login_id2
-        self.sex = packet.sex
+        self.sex = Int(packet.sex)
     }
 
     mutating func update(accountID: UInt32) {
